[Fix] Adds autofocus to the title field
########72 characters##################################################
Problem
Autofocus was being hogged by the search input. If there is another
form in the field, we want the focus to be on the first important
field of that form.

Solution
# Solution or List of Changes
Adds the autofocus attribute to the title field of the feedback form.
WIP #34

Note
# Special instructions, testing steps, rake, etc
More work needed on the search input.