[Feat]: Adds proper ordering of search results
Problem
Search results were segmented into huts and then places, ordered
alphabetically within those categories. This was ugly and a bit
unnecessary.

Solution
Learn how to implement a custom sortBy function and concat the two
search result collections together.
Closes #40 for now.

Note
This will need to be improved in the future when campsites are added!